+++ Assets/Runtime/Scripts/Engine/Modules/Rpg/RpgCharacterModifier.cs
@@ -13,7 +13,7 @@ namespace io.github.thisisnozaku.idle.framework.Engine.Modules.Rpg
     {
         public Dictionary<string, List<string>> EventTriggers;
         public string Description;
-        public RpgCharacterModifier(IdleEngine engine, long id, string description, Dictionary<string, Tuple<string, string>> modifications, Dictionary<string, List<string>> events = null) : base(engine, id, modifications)
+        public RpgCharacterModifier(IdleEngine engine, double id, string description, Dictionary<string, Tuple<string, string>> modifications, Dictionary<string, List<string>> events = null) : base(engine, id, modifications)
         {
             this.Description = description;
             this.EventTriggers = events;
